Banner story with typescript (#739)
* 🚚 Rename to .tsx

* ➕ Add types for styled components and react

* 📝 Write the story with Typescript

* ➖ Remove types as it was suddenly a separate PR

* 🎨 Get rid of types not found warning

It seems like this way of exporting things will make the warning for types not found in Storybook server go away

* 🎨 Export all banner props
