+++ apps/storybook-react/package.json
@@ -34,9 +34,6 @@
     "@storybook/addons": "^6.0.27",
     "@storybook/source-loader": "^6.0.27",
     "@storybook/theming": "^6.0.27",
-    "@types/react": "^16.9.53",
-    "@types/react-dom": "^16.9.8",
-    "@types/styled-components": "^5.1.4",
     "babel-loader": "^8.1.0",
     "react": "^16.13.1",
     "react-is": "^16.13.1"
